Updated Rule calculation algorithm
